move no-unresolved-refs to common
