+++ packages/core/src/rules/common/__tests__/fixtures/invalid-yaml.yaml
