+++ packages/core/src/rules/common/__tests__/fixtures/ref.yaml
