+++ packages/core/src/rules/overlay1/index.ts
@@ -1,7 +1,7 @@
 import { Assertions } from '../common/assertions/index.js';
 import { InfoContact } from '../common/info-contact.js';
 import { Struct } from '../common/struct.js';
-import { NoUnresolvedRefs } from '../no-unresolved-refs.js';
+import { NoUnresolvedRefs } from '../common/no-unresolved-refs.js';
 
 import type { Overlay1RuleSet } from '../../oas-types.js';
 import type { Overlay1Rule } from '../../visitors.js';
